Extend proposal and proposal vote
