+++ decidim-proposals/app/models/decidim/proposals/proposal.rb
@@ -18,11 +18,6 @@ module Decidim
       feature_manifest_name "proposals"
 
       has_many :votes, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy, counter_cache: "proposal_votes_count"
-      # Votes weight
-      has_many :up_votes, -> { where(weight: 1) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
-      has_many :down_votes, -> { where(weight: -1) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
-      has_many :neutral_votes, -> { where(weight: 0) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
-
       has_many :notes, foreign_key: "decidim_proposal_id", class_name: "ProposalNote", dependent: :destroy, counter_cache: "proposal_notes_count"
 
       validates :title, :body, presence: true
@@ -47,18 +42,6 @@ module Decidim
       # Returns Boolean.
       def voted_by?(user)
         votes.where(author: user).any?
-      end
-
-      def up_voted_by?(user)
-        votes.where(author: user,  proposal: self, weight: 1).any?
-      end
-
-      def neutral_voted_by?(user)
-        votes.where(author: user,  proposal: self, weight: 0).any?
-      end
-
-      def down_voted_by?(user)
-        votes.where(author: user,  proposal: self, weight: -1).any?
       end
 
       # Public: Checks if the organization has given an answer for the proposal.
