+++ decidim-proposals/config/initializers/extends.rb
@@ -1,5 +1,6 @@
-require "decidim/extends/proposal_extend.rb"
+require "decidim/extends/models/proposal_extend.rb"
 require "decidim/extends/commands/vote_proposal_extend.rb"
 require "decidim/extends/models/abilities/current_user_ability_extend.rb"
+require "decidim/extends/models/proposal_vote_extend.rb"
 # require "decidim/extends/controllers/concerns/orderable_extend.rb"
 # require "decidim/extends/controllers/proposal_votes_controller_extend.rb"
