+++ decidim-proposals/lib/decidim/extends/models/proposal_extend.rb
@@ -26,10 +26,27 @@ module ProposalExtend
       moderation.authorize!
     end
   end
+
+  def up_voted_by?(user)
+    votes.where(author: user,  proposal: self, weight: 1).any?
+  end
+
+  def neutral_voted_by?(user)
+    votes.where(author: user,  proposal: self, weight: 0).any?
+  end
+
+  def down_voted_by?(user)
+    votes.where(author: user,  proposal: self, weight: -1).any?
+  end
 end
 
 Decidim::Proposals::Proposal.class_eval do
   prepend(ProposalExtend)
+  # Votes weight
+  has_many :up_votes, -> { where(weight: 1) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
+  has_many :down_votes, -> { where(weight: -1) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
+  has_many :neutral_votes, -> { where(weight: 0) }, foreign_key: "decidim_proposal_id", class_name: "ProposalVote", dependent: :destroy
+
   after_create :create_proposal_moderation
   after_create :update_moderation
 end
