+++ decidim-proposals/lib/decidim/extends/models/proposal_vote_extend.rb
@@ -1,0 +1,9 @@
+Decidim::Proposals::ProposalVote.class_eval do
+  def self.create_or_delete(proposal, current_user, weight)
+    if where(author: current_user, proposal: proposal, weight: weight).any?
+      :delete
+    else
+      :post
+    end
+  end
+end
